collec

+++ MonitoringGrid.Api/CQRS/Commands/Indicator/CreateIndicatorCommand.cs
@@ -22,7 +22,7 @@ public class CreateIndicatorCommand : IRequest<Result<IndicatorDto>>
     public string? IndicatorDesc { get; set; }
 
     [Required]
-    public int CollectorID { get; set; }
+    public long CollectorID { get; set; }
 
     [Required]
     [MaxLength(255)]
